Breakapart Popover
Introduce domReadyProps pattern from useDialog
